pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tell-cli"
version = "0.1.0"
description = "Linux-only CLI that turns natural language into shell commands using Groq."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0",
  "groq>=0.12.0",
  "python-dotenv>=1.0",
]

[project.scripts]
tell = "tell.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["tell"]
